rename 268_2 keyboard files
