+++ keyboards/noxary/268_2/268_2.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "n268_2.h"
+#include "268_2.h"
 
 void matrix_init_kb(void) {
 	// put your keyboard start-up code here
